Adding comment formatting option to the nopaste plugin
This patch adds the ability to configure the format of the session as
is sent to the nopaste server. Formerly, the perl commands were
commented out and the output of them was not. This patch provides a
configuration option to swtich that behavior so that by default the
perl code is not commented out, but the output from each statement is
and this can be optionally switched back to the former behavior.


git-svn-id: http://dev.catalyst.perl.org/repos/bast/trunk/Devel-REPL@5801 bd8105ee-0ff8-0310-8827-fb3f25b6796d