Add a bit more description of Turtle plugin.
This description of what the Turtle plugin is looking for and
the default_command_prefix character value of '#' now makes it
clear that other uses of '#' may conflict with this plugin.
This is relevant to PDL shell development which uses '#' as the
shell escape character which conflicts with Turtle usages.