Update Changes and README
Note that README is just the output of 'perldoc -t Devel::REPL'
so I that POD was updated to be consistent with the information
in README and then README was generated with perldoc.  If there
is no other information in README, maybe it should be generated
by the distribution process.